test/com/tw/tictactoe/BoardTest.java
package com.tw.tictactoe;

import org.junit.Test;

import java.io.PrintStream;

import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.verify;

/**
 * Created by manderso on 5/1/14.
 */
public class BoardTest {

    @Test
    public void shouldDrawABoard() {

        PrintStream printStream = mock(PrintStream.class);
        Board board = new Board(printStream);
        board.displayBoard();
        verify(printStream).println(
                    " | | " + "\n"
                +   "-----" + "\n"
                +   " | | " + "\n"
                +   "-----" + "\n"
                +   " | | "
        );

    }
}
